Make percentage and int and wrap long comment

+++ battery.py
@@ -14,7 +14,9 @@ if not status:
     fulltext = "<span color='red'><span font='FontAwesome'>\uf00d \uf240</span></span>"
     percentleft = 100
 else:
-    # if there is more than one battery in one laptop, the percentage left is available for each battery separatelly, although state and remaining time for overall block is shown in the status of the first battery 
+    # if there is more than one battery in one laptop, the percentage left is 
+    # available for each battery separately, although state and remaining 
+    # time for overall block is shown in the status of the first battery 
     batteries = status.split("\n")
     state_batteries=[]
     commasplitstatus_batteries=[]
@@ -27,7 +29,7 @@ else:
             commasplitstatus_batteries.append(commasplitstatus)
     state = state_batteries[0]
     commasplitstatus = commasplitstatus_batteries[0]
-    percentleft = sum(percentleft_batteries) / len(percentleft_batteries)
+    percentleft = int(sum(percentleft_batteries)/len(percentleft_batteries))
     # stands for charging
     FA_LIGHTNING = "<span color='yellow'><span font='FontAwesome'>\uf0e7</span></span>"
